requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
pydantic==2.9.2
openai==1.44.0
python-multipart==0.0.9
PyYAML==6.0.2
